feat: add join! macro for structured composition of futures.
Note that unlike other join! macro implementations this does not poll unwoken futures.
It works in a similar way to tokio's JoinSet, except it allocates a single waker for all futures.
Currently it supports up to 32 futures, but more may be added in the future.

TODO: move the macros module to _priv.rs